IGNITE-20449 Remove code duplicates
